requirements.txt
fastapi>=0.95.2
moviepy>=1.0.3
openai-whisper>=20230314
python-multipart>=0.0.6
uvicorn>=0.22.0